baofeng_common.py: relocate validate_memory() - fixes 10367 Move validate_memory() to special subclasses that need to use it. GMRS-V2 (gmrsv2.py) UV-9G (baofeng_wp970i.py)
